yolo-server/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
ultralytics==8.3.0
python-multipart==0.0.12
Pillow==11.0.0
numpy==2.1.0
requests==2.32.0
opencv-python-headless==4.10.0.84
